Add static cast in examples

+++ examples/AWS_IoTCore/AWS_IoTCore.ino
@@ -507,7 +507,7 @@ bool setupModem() {
 
 #if TINY_GSM_USE_GPRS
   // Unlock your SIM card with a PIN if needed
-  if (GSM_PIN && modem.getSimStatus() != 3) {
+  if (GSM_PIN && static_cast<int>(modem.getSimStatus()) != 3) {
     success &= modem.simUnlock(GSM_PIN);
   }
 #endif
